+++ CMakeLists.txt
@@ -42,18 +42,17 @@ target_include_directories(promise PUBLIC include)
 
 if(PROMISE_BUILD_EXAMPLES) 
     add_library(common_example_options INTERFACE)
-    target_include_directories(common_example_options INTERFACE .)
 endif()
 
 find_package(Boost)
 if(NOT Boost_FOUND)
     message(WARNING "Boost not found, so asio projects will not be compiled")
 else()
-    target_include_directories(common_example_options INTERFACE ${Boost_INCLUDE_DIRS})
+    target_include_directories(promise PUBLIC ${Boost_INCLUDE_DIRS})
     if(UNIX)
-        target_link_libraries(common_example_options INTERFACE pthread)
+        target_link_libraries(promise PUBLIC pthread)
     elseif(WIN32)
-        target_link_libraries(common_example_options INTERFACE -lws2_32 -lmswsock)
+        target_link_libraries(promise PUBLIC -lws2_32 -lmswsock)
     endif()
 endif()
 
@@ -67,7 +66,7 @@ else()
     else()
         add_library(promise_qt STATIC ./add_ons/qt/promise_qt.cpp ${my_headers})
     endif()
-    target_link_libraries(promise_qt PRIVATE promise Qt${QT_VERSION_MAJOR}::Widgets)
+    target_link_libraries(promise_qt PUBLIC promise Qt${QT_VERSION_MAJOR}::Widgets)
 endif()
 
 if(PROMISE_BUILD_EXAMPLES)
