--- add_ons/qt/promise_qt.cpp
@@ -1,5 +1,0 @@
-
-#ifndef PROMISE_HEADONLY
-#include "promise_qt_inl.hpp"
-#endif
-
