+++ include/promise-cpp/add_ons/qt/promise_qt.cpp
@@ -1,0 +1,5 @@
+
+#ifndef PROMISE_HEADONLY
+#include "promise_qt_inl.hpp"
+#endif
+
